Calorie Negation basic output implemented
The data method for Calorie Negation is now in place. The output isn't
pretty, but the data seems to be accurate (though the BMR input is
static currently).

Also, I think the method of getting MET values needs to be changed. I'm
using local variables here, and I think, in the future, it should be
method calls to return unique MET values instead of discrete levels of
perceived exertion.