Calorie Negation stores workouts to DB
Calorie Negation now displays types of workouts along with needed times
and add buttons. Hitting an add button will return the user to the home
screen. If you check the debug menu, you will see that it has
successfully added. Showing up on a schedule will be coming later this
morning.

Also, I moved the minimum required API level back to 15. If the only
place that was needed was Calorie Negation's "setBackground" calls, I
did some research, and there are corresponding deprecated calls that can
be used. It's a little messy, but I don't want to have to change our
stated minimum unless we see something for which there is zero
workaround.